# openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Tenant OpenSSF Scan API
  version: 1.0.0
  description: Synchronous, fixture-backed OpenSSF Scorecard scan API with static bearer authentication and tenant isolation.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
tags:
  - name: scans
    description: Scan creation and retrieval
paths:
  /scans:
    post:
      tags: [scans]
      summary: Trigger a scan for a repository URL
      operationId: createScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
      responses:
        '201':
          description: Scan created synchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /scans/{scan_id}:
    get:
      tags: [scans]
      summary: Get scan metadata and summary
      operationId: getScan
      parameters:
        - $ref: '#/components/parameters/ScanId'
      responses:
        '200':
          description: Scan metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /scans/{scan_id}/issues:
    get:
      tags: [scans]
      summary: List normalized security issues for a scan
      operationId: listScanIssues
      parameters:
        - $ref: '#/components/parameters/ScanId'
        - name: min_severity
          in: query
          required: false
          description: Inclusive minimum severity. Severity order is LOW < MEDIUM < HIGH < CRITICAL.
          schema:
            $ref: '#/components/schemas/Severity'
      responses:
        '200':
          description: Normalized issue list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: static-token
  parameters:
    ScanId:
      name: scan_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        pattern: '^scn_[A-Za-z0-9_-]+$'
  schemas:
    Severity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    ScanRequest:
      type: object
      required: [repository_url]
      additionalProperties: false
      properties:
        repository_url:
          type: string
          format: uri
          pattern: '^https://(github\.com|gitlab\.com)/[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+/?$'
          description: HTTPS GitHub or GitLab repository URL.
          example: https://github.com/acme-corp/payment-gateway
    Scan:
      type: object
      required: [scan_id, tenant_id, repository_url, status, aggregate_score, created_at]
      additionalProperties: false
      properties:
        scan_id:
          type: string
          pattern: '^scn_[A-Za-z0-9_-]+$'
        tenant_id:
          type: string
          minLength: 1
        repository_url:
          type: string
          format: uri
        status:
          type: string
          enum: [COMPLETED]
        aggregate_score:
          type: number
          format: float
          minimum: 0
          maximum: 10
        created_at:
          type: string
          format: date-time
    SecurityIssue:
      type: object
      required: [issue_id, check_name, score, severity, title, description, evidence, remediation_hint]
      additionalProperties: false
      properties:
        issue_id:
          type: string
        check_name:
          type: string
          minLength: 1
        score:
          type: integer
          minimum: 0
          maximum: 10
        severity:
          $ref: '#/components/schemas/Severity'
        title:
          type: string
        description:
          type: string
        evidence:
          type: array
          items:
            type: string
        remediation_hint:
          type: string
    IssueList:
      type: object
      required: [scan_id, issues]
      additionalProperties: false
      properties:
        scan_id:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/SecurityIssue'
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                required: [field, issue]
                additionalProperties: false
                properties:
                  field: {type: string}
                  issue: {type: string}
  responses:
    BadRequest:
      description: Malformed request or invalid query parameter
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Missing or unknown bearer token
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: Authenticated user is not permitted to perform the operation
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Scan does not exist or belongs to another tenant
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    InternalError:
      description: Unexpected fixture/storage failure
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}